check if file exists

+++ MITADS/utils/line_rules.py
@@ -1,5 +1,6 @@
 import re
 import json
+import os
 valid_char_regex = re.compile(r"[^\s'abcdefghijklmnopqrstuvwxyzàèéìíòóôùúABCDEFGHIJKLMNOPQRSTUVWXYZÀÈÉÌÍÒÓÔÙÚ,.!?:;]")
 
 
@@ -21,54 +22,55 @@ class LineRules:
             self.discarded.setdefault(rule, []).append(line)
 
     def save(self):
-        with open(self.discard_file, 'w') as f:
-            json.dump(self.discarded, f)
+        if os.path.exists(self.discard_file):
+            with open(self.discard_file, 'w') as f:
+                json.dump(self.discarded, f)
 
     def is_not_valid(self, line):
         not_valid = valid_char_regex.search(line)
         if not_valid: self.discard("is_not_valid", line)
         return not_valid
-    
+
     def startswith(self, line, chars):
         for char in chars:
             if line.startswith(char):
                 self.discard("startswith_" + str(chars) , line)
                 return True
-            
+
     def endswith(self, line, chars):
         for char in chars:
             if line.endswith(char):
                 self.discard("endswith_" + str(chars), line)
                 return True
-    
+
     def contain(self, line, chars):
         for char in chars:
             if line.find(char) >= 0:
                 self.discard("contain" + str(chars), line)
                 return True
-    
+
     def isdigit(self, searchme):
         for search in searchme:
             if search.isdigit():
                 self.discard("is_digit", search)
                 return True
-            
+
     def isbookref(self, text):
         text = text.replace('.', '')
         if text.find(',') >= 2 and text[-2:].isdigit():
             self.discard("isbookref", text)
             return True
-        
+
     def isbrokensimplebracket(self, text):
         if text.find('(') >= 1 and text.find(')') == -1:
             self.discard("isbrokensimplebracket", text)
             return True
-        
+
     def isbrokenparenthesis(self, text):
         if text.find(')') >= 1 and text.find('(') == -1 or text.find('(') >= 1 and text.find(')') == -1:
             self.discard("isbrokenparenthesis", text)
             return True
-        
+
     def hasafinalrepeated(self, text):
         for word in text.split():
             if len(word) >= 2 and not word.strip().isnumeric():
@@ -77,7 +79,7 @@ class LineRules:
                     self.discard("hasafinalrepeated", text)
                     return True
         return False
-    
+
     def parenthesismatch(self, text):
         last = None
         for char in text:
